requirements.txt
pandas
numpy
pyarrow
scikit-learn
scipy
plotly
streamlit
pyyaml
ruff
black
isort
pytest
joblib
rich
tqdm
